+++ CHANGELOG.rst
@@ -2,13 +2,18 @@
 Changelog
 =========
 
+
 x.x.x ?
-==================
+=======
 
 * AlertRulev9 propagates dashbord_uid and panel_uid to the alert rule
+* Added ``tooltipSort`` parameter to PieChartv2 panel
+* Fix mappings for Table
+* Added support for AWS Cross-Account in CloudwatchMetricsTarget
+* Added `LokiTarget`
 
 0.7.1 2024-01-12
-==================
+================
 
 * Extended DashboardLink to support links to dashboards and urls, as per the docs_